@@ -35,22 +35,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.0.0",
-    "@angular/cdk": "^12.0.1",
-    "@angular/cdk-experimental": "^12.0.1",
-    "@angular/common": "^12.0.0",
-    "@angular/compiler": "^12.0.0",
-    "@angular/components-examples": "angular/material2-docs-content#12.0.x",
-    "@angular/core": "^12.0.0",
-    "@angular/forms": "^12.0.0",
-    "@angular/google-maps": "^12.0.1",
-    "@angular/material": "^12.0.1",
-    "@angular/material-experimental": "^12.0.1",
-    "@angular/material-moment-adapter": "^12.0.1",
-    "@angular/platform-browser": "^12.0.0",
-    "@angular/platform-browser-dynamic": "^12.0.0",
-    "@angular/router": "^12.0.0",
-    "@angular/youtube-player": "^12.0.1",
+    "@angular/animations": "^12.1.0",
+    "@angular/cdk": "^12.1.1",
+    "@angular/cdk-experimental": "^12.1.1",
+    "@angular/common": "^12.1.0",
+    "@angular/compiler": "^12.1.0",
+    "@angular/components-examples": "angular/material2-docs-content#12.1.x",
+    "@angular/core": "^12.1.0",
+    "@angular/forms": "^12.1.0",
+    "@angular/google-maps": "^12.1.1",
+    "@angular/localize": "^12.1.0",
+    "@angular/material": "^12.1.1",
+    "@angular/material-experimental": "^12.1.1",
+    "@angular/material-moment-adapter": "^12.1.1",
+    "@angular/platform-browser": "^12.1.0",
+    "@angular/platform-browser-dynamic": "^12.1.0",
+    "@angular/router": "^12.1.0",
+    "@angular/youtube-player": "^12.1.1",
     "material-components-web": "12.0.0-canary.b52196498.0",
     "moment": "^2.29.1",
     "rxjs": "^6.6.7",
@@ -58,15 +59,14 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.0.0",
-    "@angular-eslint/builder": "4.0.0",
-    "@angular-eslint/eslint-plugin": "4.0.0",
-    "@angular-eslint/eslint-plugin-template": "4.0.0",
-    "@angular-eslint/schematics": "4.0.0",
-    "@angular-eslint/template-parser": "4.0.0",
-    "@angular/cli": "^12.0.0",
-    "@angular/compiler-cli": "^12.0.0",
-    "@angular/localize": "^12.0.0",
+    "@angular-devkit/build-angular": "^12.1.0",
+    "@angular-eslint/builder": "^12.1.0",
+    "@angular-eslint/eslint-plugin": "^12.1.0",
+    "@angular-eslint/eslint-plugin-template": "^12.1.0",
+    "@angular-eslint/schematics": "^12.1.0",
+    "@angular-eslint/template-parser": "^12.1.0",
+    "@angular/cli": "^12.1.0",
+    "@angular/compiler-cli": "^12.1.0",
     "@types/imagemin": "^7.0.0",
     "@types/jasmine": "^3.6.6",
     "@types/node": "^14.14.22",
